requirements.txt
anthropic>=0.40.0
openai>=1.50.0
rich>=13.7.0
click>=8.1.0
jinja2>=3.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pyyaml>=6.0.0
weasyprint>=62.0
xmltodict>=0.13.0
python-nmap>=0.7.1
dnspython>=2.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
